backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-cloud-aiplatform==1.38.1
gunicorn==21.2.0